Added time into trade history (and changed it from size_t to uint64 in trade struct) also implemented cancel_order into order book class.

+++ include/trade.hpp
@@ -1,10 +1,11 @@
 #pragma once
 #include "order.hpp"
 #include <cstddef>
+#include <cstdint>
 
 struct Trade 
 {
-    size_t time;
+    uint64_t time;
     size_t size;
     long price;
     Order_type type;
